implemented illust detection
