+++ figma/src/flutter/index.ts
@@ -16,7 +16,9 @@ import { makeDivider } from "./make/divider.make";
 import { detectIfButton } from "@reflect.bridged.xyz/detection/lib/button.detection";
 import { makeButton } from "./make/button.make";
 import { detectIfIcon } from "@reflect.bridged.xyz/detection/lib/icon.detection";
-import { makeMaterialIcon } from "./make/icon.make";
+import { makeMaterialIcon, makePlaceholderIcon } from "./make/icon.make";
+import { detectIfIllust } from "@reflect.bridged.xyz/detection/lib/illust.detection";
+import { makePlaceHolderImage } from "./make/image.make";
 
 
 let parentId = "";
@@ -103,7 +105,13 @@ function flutterWidgetGenerator(sceneNode: ReadonlyArray<ReflectSceneNode> | Ref
     const iconDetectionResult = detectIfIcon(node)
     if (iconDetectionResult.result) {
       console.log('this node is detected as an icon.', node.name)
-      return makeMaterialIcon('add')
+      return makePlaceholderIcon()
+    }
+
+    const illustDetectionResult = detectIfIllust(node)
+    if (illustDetectionResult.result) {
+      console.log('this node is detected as an illust.', node.name)
+      return makePlaceHolderImage(node)
     }
 
     // console.log(`starting handling node of ${node.name} type of ${node.type}`)
