+++ figma/src/utils/color.ts
@@ -1,21 +1,3 @@
-export function rgbTo6hex(color: RGB | RGBA): string {
-  const hex = ((color.r * 255) | (1 << 8)).toString(16).slice(1) +
-    ((color.g * 255) | (1 << 8)).toString(16).slice(1) +
-    ((color.b * 255) | (1 << 8)).toString(16).slice(1);
-
-  return hex;
-}
-
-export function rgbTo8hex(color: RGB, alpha: number): string {
-  // when color is RGBA, alpha is set automatically
-  // when color is RGB, alpha need to be set manually (default: 1.0)
-  const hex = ((alpha * 255) | (1 << 8)).toString(16).slice(1) +
-    ((color.r * 255) | (1 << 8)).toString(16).slice(1) +
-    ((color.g * 255) | (1 << 8)).toString(16).slice(1) +
-    ((color.b * 255) | (1 << 8)).toString(16).slice(1);
-
-  return hex;
-}
 
 export function gradientAngle(fill: GradientPaint): number {
   const decomposed = decomposeRelativeTransform(
