+++ packages/uiutils/lib/index.ts
@@ -1,1 +1,2 @@
 export * from "./pixels"
+export * from "./colors"
